+++ spi/src/lib.rs
@@ -13,25 +13,22 @@ extern crate alloc;
 #[cfg(feature = "async")]
 pub mod asynch;
 
-use embedded_hal::{digital::OutputPin, spi::SpiBusWrite};
-
+use byte_slice_cast::*;
 use display_interface::{DataFormat, DisplayError, WriteOnlyDataCommand};
+use embedded_hal::{digital::OutputPin, spi::SpiDeviceWrite};
 
 type Result = core::result::Result<(), DisplayError>;
 
 fn send_u8<SPI>(spi: &mut SPI, words: DataFormat<'_>) -> Result
 where
-    SPI: SpiBusWrite,
+    SPI: SpiDeviceWrite,
 {
     match words {
         DataFormat::U8(slice) => spi.write(slice).map_err(|_| DisplayError::BusWriteError),
-        DataFormat::U16(slice) => {
-            use byte_slice_cast::*;
-            spi.write(slice.as_byte_slice())
-                .map_err(|_| DisplayError::BusWriteError)
-        }
+        DataFormat::U16(slice) => spi
+            .write(slice.as_byte_slice())
+            .map_err(|_| DisplayError::BusWriteError),
         DataFormat::U16LE(slice) => {
-            use byte_slice_cast::*;
             for v in slice.as_mut() {
                 *v = v.to_le();
             }
@@ -39,7 +36,6 @@ where
                 .map_err(|_| DisplayError::BusWriteError)
         }
         DataFormat::U16BE(slice) => {
-            use byte_slice_cast::*;
             for v in slice.as_mut() {
                 *v = v.to_be();
             }
@@ -68,7 +64,6 @@ where
             Ok(())
         }
         DataFormat::U16LEIter(iter) => {
-            use byte_slice_cast::*;
             let mut buf = [0; 32];
             let mut i = 0;
 
@@ -91,7 +86,6 @@ where
             Ok(())
         }
         DataFormat::U16BEIter(iter) => {
-            use byte_slice_cast::*;
             let mut buf = [0; 64];
             let mut i = 0;
             let len = buf.len();
@@ -120,87 +114,15 @@ where
 
 /// SPI display interface.
 ///
-/// This combines the SPI peripheral and a data/command as well as a chip-select pin
-pub struct SPIInterface<SPI, DC, CS> {
-    spi_no_cs: SPIInterfaceNoCS<SPI, DC>,
-    cs: CS,
-}
-
-impl<SPI, DC, CS> SPIInterface<SPI, DC, CS>
-where
-    CS: OutputPin,
-{
-    pub(crate) fn cs_low(&mut self) -> Result {
-        self.cs.set_low().map_err(|_| DisplayError::CSError)?;
-        Ok(())
-    }
-
-    pub(crate) fn cs_high(&mut self) -> Result {
-        self.cs.set_high().map_err(|_| DisplayError::CSError)?;
-        Ok(())
-    }
-}
-
-impl<SPI, DC, CS> SPIInterface<SPI, DC, CS>
-where
-    SPI: SpiBusWrite,
-    DC: OutputPin,
-    CS: OutputPin,
-{
-    /// Create new SPI interface for communication with a display driver
-    pub fn new(spi: SPI, dc: DC, cs: CS) -> Self {
-        Self {
-            spi_no_cs: SPIInterfaceNoCS::new(spi, dc),
-            cs,
-        }
-    }
-
-    /// Consume the display interface and return
-    /// the underlying peripherial driver and GPIO pins used by it
-    pub fn release(self) -> (SPI, DC, CS) {
-        let (spi, dc) = self.spi_no_cs.release();
-        (spi, dc, self.cs)
-    }
-
-    fn with_cs(&mut self, f: impl FnOnce(&mut SPIInterfaceNoCS<SPI, DC>) -> Result) -> Result {
-        // Assert chip select pin
-        self.cs_low().map_err(|_| DisplayError::CSError)?;
-
-        let result = f(&mut self.spi_no_cs);
-
-        // Deassert chip select pin
-        self.cs_high().ok();
-
-        result
-    }
-}
-
-impl<SPI, DC, CS> WriteOnlyDataCommand for SPIInterface<SPI, DC, CS>
-where
-    SPI: SpiBusWrite,
-    DC: OutputPin,
-    CS: OutputPin,
-{
-    fn send_commands(&mut self, cmds: DataFormat<'_>) -> Result {
-        self.with_cs(|spi_no_cs| spi_no_cs.send_commands(cmds))
-    }
-
-    fn send_data(&mut self, buf: DataFormat<'_>) -> Result {
-        self.with_cs(|spi_no_cs| spi_no_cs.send_data(buf))
-    }
-}
-
-/// SPI display interface.
-///
 /// This combines the SPI peripheral and a data/command pin
-pub struct SPIInterfaceNoCS<SPI, DC> {
+pub struct SPIInterface<SPI, DC> {
     spi: SPI,
     dc: DC,
 }
 
-impl<SPI, DC> SPIInterfaceNoCS<SPI, DC>
+impl<SPI, DC> SPIInterface<SPI, DC>
 where
-    SPI: SpiBusWrite,
+    SPI: SpiDeviceWrite,
     DC: OutputPin,
 {
     /// Create new SPI interface for communication with a display driver
@@ -215,9 +137,9 @@ where
     }
 }
 
-impl<SPI, DC> WriteOnlyDataCommand for SPIInterfaceNoCS<SPI, DC>
+impl<SPI, DC> WriteOnlyDataCommand for SPIInterface<SPI, DC>
 where
-    SPI: SpiBusWrite,
+    SPI: SpiDeviceWrite,
     DC: OutputPin,
 {
     fn send_commands(&mut self, cmds: DataFormat<'_>) -> Result {